refactor: IPNE 描画 effect 本体を renderGameFrame へ単一抽出
- Game.tsx の描画ロジック（645行）を renderGameFrame(rc) へ逐語移植し effect を薄く
- 描画順序・依存配列は不変。後続で層へ分割する土台
- Game.tsx: 990行 → 262行。renderGameFrame.ts: 800行
